4.2.1. Небольшой рефакторинг
